chore: move burst back to 5
