+++ internal/smf/util/flow_desc.go
@@ -2,7 +2,7 @@
 // Copyright 2019 Communication Service/Software Laboratory, National Chiao Tung University (free5gc.org)
 // SPDX-License-Identifier: Apache-2.0
 
-package flowdesc
+package util
 
 import (
 	"fmt"
